Error due to l10n exports with default field.
Hello!
We have an error due to this part of code because it is outside the component and after build it causes error because this code use global l10n module in bundle with nested `default` filed. That is we should apply it like `l10n.default.en` and etc or replace this code into component module.
Our team would be eternally grateful if you approve this pull request!